requirements.txt
flask
PySerial